backend2/requirements.txt
# Core PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Swin Transformer and Vision Models
timm>=0.9.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Scientific Computing
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars and Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# Configuration and Utilities
PyYAML>=6.0
omegaconf>=2.3.0

# Video Processing
av>=10.0.0  # PyAV for torchvision.io.read_video
moviepy>=1.0.3
imageio>=2.31.0

# GPU Acceleration (optional but recommended)
# Note: Install CUDA version matching your PyTorch installation
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for experimentation (optional)
jupyter>=1.0.0
ipykernel>=6.23.0

# Research Paper Dependencies
# For Springer conference publication
reportlab>=4.0.0  # PDF generation
jupyter-book>=0.15.0  # Documentation
sphinx>=7.0.0  # Documentation
sphinx-rtd-theme>=1.3.0  # Read the Docs theme

